Fix view_count overflow on high-view videos (e.g. Gangnam Style at 6B+ views)
Found while testing the new /summarize/queue endpoint: Postgres INTEGER
tops out around 2.1B, and several real videos exceed that view count.
The insert was failing with "integer out of range" inside the background
task, silently (the task has no request left to report the failure to —
which is itself a case where a wide-view video summarized via the queue
endpoint would just never show up, with nothing in the response to
explain why).

Changed Summary.view_count to BigInteger and taught the existing
additive-only auto-migration to also widen INTEGER columns to BIGINT
when the model calls for it — widening never loses data, so it's safe to
run unattended the same way column additions already are.

+++ app/models.py
@@ -2,7 +2,7 @@
 
 from datetime import datetime, timezone
 
-from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, String, Text
+from sqlalchemy import BigInteger, Boolean, DateTime, ForeignKey, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.db import Base
@@ -59,7 +59,7 @@ class Summary(Base):
     channel: Mapped[str | None] = mapped_column(String(255), nullable=True)
     channel_id: Mapped[str | None] = mapped_column(String(64), nullable=True)
     published_at: Mapped[str | None] = mapped_column(String(32), nullable=True)
-    view_count: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    view_count: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     comment_count: Mapped[int | None] = mapped_column(Integer, nullable=True)
     like_count: Mapped[int | None] = mapped_column(Integer, nullable=True)
     duration_seconds: Mapped[int | None] = mapped_column(Integer, nullable=True)
